SVN r1695 (odameXbox)

+++ common/r_defs.h
@@ -386,9 +386,9 @@ struct drawseg_s
     
     // Pointers to lists for sprite clipping,
     //  all three adjusted so [x1] is first value.
-    short*		sprtopclip;		
-    short*		sprbottomclip;	
-    short*		maskedtexturecol;
+    int*		sprtopclip;		
+    int*		sprbottomclip;	
+    int*		maskedtexturecol;
 };
 typedef struct drawseg_s drawseg_t;
 
@@ -536,9 +536,9 @@ struct visplane_s
 	fixed_t		xscale, yscale;		// [RH] Support flat scaling
 	angle_t		angle;				// [RH] Support flat rotation
 
-	unsigned short *bottom;			// [RH] bottom and top arrays are dynamically
-	unsigned short pad;				//		allocated immediately after the
-	unsigned short top[3];			//		visplane.
+	unsigned int *bottom;			// [RH] bottom and top arrays are dynamically
+	unsigned int pad;				//		allocated immediately after the
+	unsigned int top[3];			//		visplane.
 };
 typedef struct visplane_s visplane_t;
